requirements.txt
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
greenlet==1.1.3
idna==3.4
lxml==4.9.1
psycopg2-binary==2.9.3
requests==2.28.1
soupsieve==2.3.2.post1
SQLAlchemy==1.4.41
urllib3==1.26.12
